requirements.txt
msal==1.31.1
requests==2.32.3
openpyxl==3.1.5
python-dotenv==1.0.1
